requirements.txt
Pillow
django-tables2
django-widget-tweaks
